added list for commandtexts and created commandtext.cs, commandcreator.cs and renamed the old commandcreator.cs to commandcreatorplugin.cs
