+++ iBot/Plugins/CommandCreator/CommandText.cs
@@ -1,0 +1,21 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace IBot.Plugins.CommandCreator
+{
+    internal class CommandText
+    {
+        public CommandText(int id, string name)
+        {
+            Id = id;
+            Name = name;
+        }
+
+        public int Id { get; }
+
+        public string Name { get; }
+    }
+}
